requirements.txt
flask
python-dotenv
sqlalchemy
pymysql
lancedb
sentence-transformers
torch
litellm
python-docx
notebook
python-pptx